+++ src/modules/comments/comment.service.ts
@@ -1,4 +1,5 @@
 import commentRepository from "./comment.repository";
+import cardService from "../cards/card.service";
 import { Comment } from "../../database/entities/comment";
 import customError from "../../common/error/customError";
 import { Result } from "../../common/response/Result";
@@ -6,40 +7,63 @@ import { Result } from "../../common/response/Result";
 class CommentService {
     public async findCommentById(id: number): Promise<Result> {
         try {
+            if (!id) {
+                throw new customError(400, "Id is required");
+            }
             const comment = await commentRepository.findCommentById(id);
             if (!comment) {
                 throw new customError(404, "Comment not found");
             }
             return new Result(true, 200, "Comment found", comment);
         } catch (error) {
-            throw new customError(400, `CommentService has error: ${error}`);
+            throw error;
         }
     }
 
-    public async createComment(comment: Comment): Promise<Result> {
+    
+
+    public async createComment(comment: Comment, cardId: number): Promise<Result> {
         try {
+            if (!cardId) {
+                throw new customError(400, "Card id is required");
+            }
             const newComment = await commentRepository.createComment(comment);
+            await cardService.addCommentToCard(cardId, newComment.id);
             return new Result(true, 201, "Comment created", newComment);
         } catch (error) {
-            throw new customError(400, `CommentService has error: ${error}`);
+            throw error;
         }
     }
 
     public async deleteComment(id: number): Promise<Result> {
         try {
+            if (!id) {
+                throw new customError(400, "Id is required");
+            }
+            const commentExist = await commentRepository.findCommentById(id);
+            if (!commentExist) {
+                throw new customError(404, "Comment not found");
+            }
             await commentRepository.deleteComment(id);
-            return new Result(true, 204, "Comment deleted");
+            return new Result(true, 200, "Comment deleted");
         } catch (error) {
-            throw new customError(400, `CommentService has error: ${error}`);
+            throw error;
         }
     }
 
     public async updateComment(id: number, comment: Partial<Comment>): Promise<Result> {
         try {
+            if (!id) {
+                throw new customError(400, "Id is required");
+            }
+            const commentExist = await commentRepository.findCommentById(id);
+            if (!commentExist) {
+                throw new customError(404, "Comment not found");
+            }
             await commentRepository.updateComment(id, comment);
             return new Result(true, 200, "Comment updated");
         } catch (error) {
-            throw new customError(400, `CommentService has error: ${error}`);
+            throw error;
         }
     }
 }
